requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
llama-index-core>=0.10.0
llama-index-llms-ollama>=0.1.1
llama-index-embeddings-ollama>=0.1.1
llama-index-readers-file>=0.1.1
pypdf>=4.0.1
pydantic>=2.5.2
python-dotenv>=1.0.0
aiofiles>=23.2.1
jinja2>=3.1.3
